refactor onTeabSearchResultsAvailable with eventController
